checkpoint/7th_checkpoint/community_real/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
python-jose
passlib[bcrypt]
httpx
python-multipart
fastapi-jwt-auth
python-dotenv
aiofiles
